don't bail at the top when a track doesn't have a path, as doing this makes tracks from itunes match unplayable.
the only bits that depend on tracks having a path seem to be script actions, and it looks like those are on the chopping block already, anyhow, so i've just wrapped each thing that depends on a path in a conditional to not even try. this is a hack, but it makes things work.